requirements.txt
hydra-core>=1.2.0,<1.3
omegaconf>=2.2,<2.3
pynvml==11.4.1
requests==2.26.0
tqdm==4.62.3
zstandard==0.15.2
huggingface_hub>=0.13.0
dask
img2dataset